@@ -1,12 +1,12 @@
-const clone = require('clone')
-const debug = require('debug')('bittorrent-tracker:websocket-tracker')
-const Peer = require('simple-peer')
-const randombytes = require('randombytes')
-const Socket = require('simple-websocket')
-const Socks = require('socks')
+import Debug from 'debug'
+import Peer from '@thaunknown/simple-peer/lite.js'
+import Socket from '@thaunknown/simple-websocket'
+import { arr2text, arr2hex, hex2bin, bin2hex, randomBytes } from 'uint8-util'
 
-const common = require('../common')
-const Tracker = require('./tracker')
+import common from '../common.js'
+import Tracker from './tracker.js'
+
+const debug = Debug('bittorrent-tracker:websocket-tracker')
 
 // Use a socket pool, so tracker clients share WebSocket objects for the same server.
 // In practice, WebSockets are pretty slow to establish, so this gives a nice performance
@@ -78,8 +78,8 @@ class WebSocketTracker extends Tracker {
     }
 
     const infoHashes = (Array.isArray(opts.infoHash) && opts.infoHash.length > 0)
-      ? opts.infoHash.map(infoHash => infoHash.toString('binary'))
-      : (opts.infoHash && opts.infoHash.toString('binary')) || this.client._infoHashBinary
+      ? opts.infoHash.map(infoHash => hex2bin(infoHash))
+      : (opts.infoHash && hex2bin(opts.infoHash)) || this.client._infoHashBinary
     const params = {
       action: 'scrape',
       info_hash: infoHashes
@@ -183,7 +183,7 @@ class WebSocketTracker extends Tracker {
       if (this.client._proxyOpts) {
         agent = parsedUrl.protocol === 'wss:' ? this.client._proxyOpts.httpsAgent : this.client._proxyOpts.httpAgent
         if (!agent && this.client._proxyOpts.socksProxy) {
-          agent = new Socks.Agent(clone(this.client._proxyOpts.socksProxy), (parsedUrl.protocol === 'wss:'))
+          agent = this.client._proxyOpts.socksProxy
         }
       }
       this.socket = socketPool[this.announceUrl] = new Socket({ url: this.announceUrl, agent })
@@ -212,7 +212,7 @@ class WebSocketTracker extends Tracker {
     this.expectingResponse = false
 
     try {
-      data = JSON.parse(data)
+      data = JSON.parse(arr2text(data))
     } catch (err) {
       this.client.emit('warning', new Error('Invalid tracker response'))
       return
@@ -231,7 +231,7 @@ class WebSocketTracker extends Tracker {
     if (data.info_hash !== this.client._infoHashBinary) {
       debug(
         'ignoring websocket data from %s for %s (looking for %s: reused socket)',
-        this.announceUrl, common.binaryToHex(data.info_hash), this.client.infoHash
+        this.announceUrl, bin2hex(data.info_hash), this.client.infoHash
       )
       return
     }
@@ -264,7 +264,7 @@ class WebSocketTracker extends Tracker {
     if (data.complete != null) {
       const response = Object.assign({}, data, {
         announce: this.announceUrl,
-        infoHash: common.binaryToHex(data.info_hash)
+        infoHash: bin2hex(data.info_hash)
       })
       this.client.emit('update', response)
     }
@@ -273,7 +273,7 @@ class WebSocketTracker extends Tracker {
     if (data.offer && data.peer_id) {
       debug('creating peer (from remote offer)')
       peer = this._createPeer()
-      peer.id = common.binaryToHex(data.peer_id)
+      peer.id = bin2hex(data.peer_id)
       peer.once('signal', answer => {
         const params = {
           action: 'announce',
@@ -291,10 +291,10 @@ class WebSocketTracker extends Tracker {
     }
 
     if (data.answer && data.peer_id) {
-      const offerId = common.binaryToHex(data.offer_id)
+      const offerId = bin2hex(data.offer_id)
       peer = this.peers[offerId]
       if (peer) {
-        peer.id = common.binaryToHex(data.peer_id)
+        peer.id = bin2hex(data.peer_id)
         this.client.emit('peer', peer)
         peer.signal(data.answer)
 
@@ -321,7 +321,7 @@ class WebSocketTracker extends Tracker {
       // (separate from announce interval)
       const response = Object.assign(data[infoHash], {
         announce: this.announceUrl,
-        infoHash: common.binaryToHex(infoHash)
+        infoHash: bin2hex(infoHash)
       })
       this.client.emit('scrape', response)
     })
@@ -374,13 +374,13 @@ class WebSocketTracker extends Tracker {
     checkDone()
 
     function generateOffer () {
-      const offerId = randombytes(20).toString('hex')
+      const offerId = arr2hex(randomBytes(20))
       debug('creating peer (from _generateOffers)')
       const peer = self.peers[offerId] = self._createPeer({ initiator: true })
       peer.once('signal', offer => {
         offers.push({
           offer,
-          offer_id: common.hexToBinary(offerId)
+          offer_id: hex2bin(offerId)
         })
         checkDone()
       })
@@ -439,4 +439,4 @@ WebSocketTracker._socketPool = socketPool
 
 function noop () {}
 
-module.exports = WebSocketTracker
+export default WebSocketTracker
