@@ -1,7 +1,9 @@
-const arrayRemove = require('unordered-array-remove')
-const debug = require('debug')('bittorrent-tracker:swarm')
-const LRU = require('lru')
-const randomIterate = require('random-iterate')
+import arrayRemove from 'unordered-array-remove'
+import Debug from 'debug'
+import LRU from 'lru'
+import randomIterate from 'random-iterate'
+
+const debug = Debug('bittorrent-tracker:swarm')
 
 // Regard this as the default implementation of an interface that you
 // need to support when overriding Server.createSwarm() and Server.getSwarm()
@@ -161,4 +163,4 @@ class Swarm {
   }
 }
 
-module.exports = Swarm
+export default Swarm
