@@ -1,19 +1,21 @@
-const bencode = require('bencode')
-const debug = require('debug')('bittorrent-tracker:server')
-const dgram = require('dgram')
-const EventEmitter = require('events')
-const http = require('http')
-const peerid = require('bittorrent-peerid')
-const series = require('run-series')
-const string2compact = require('string2compact')
-const WebSocketServer = require('ws').Server
+import bencode from 'bencode'
+import Debug from 'debug'
+import dgram from 'dgram'
+import EventEmitter from 'events'
+import http from 'http'
+import peerid from 'bittorrent-peerid'
+import series from 'run-series'
+import string2compact from 'string2compact'
+import { WebSocketServer } from 'ws'
+import { hex2bin } from 'uint8-util'
 
-const common = require('./lib/common')
-const Swarm = require('./lib/server/swarm')
-const parseHttpRequest = require('./lib/server/parse-http')
-const parseUdpRequest = require('./lib/server/parse-udp')
-const parseWebSocketRequest = require('./lib/server/parse-websocket')
+import common from './lib/common.js'
+import Swarm from './lib/server/swarm.js'
+import parseHttpRequest from './lib/server/parse-http.js'
+import parseUdpRequest from './lib/server/parse-udp.js'
+import parseWebSocketRequest from './lib/server/parse-websocket.js'
 
+const debug = Debug('bittorrent-tracker:server')
 const hasOwnProperty = Object.prototype.hasOwnProperty
 
 /**
@@ -355,7 +357,7 @@ class Server extends EventEmitter {
   }
 
   createSwarm (infoHash, cb) {
-    if (Buffer.isBuffer(infoHash)) infoHash = infoHash.toString('hex')
+    if (ArrayBuffer.isView(infoHash)) infoHash = infoHash.toString('hex')
 
     process.nextTick(() => {
       const swarm = this.torrents[infoHash] = new Server.Swarm(infoHash, this)
@@ -364,7 +366,7 @@ class Server extends EventEmitter {
   }
 
   getSwarm (infoHash, cb) {
-    if (Buffer.isBuffer(infoHash)) infoHash = infoHash.toString('hex')
+    if (ArrayBuffer.isView(infoHash)) infoHash = infoHash.toString('hex')
 
     process.nextTick(() => {
       cb(null, this.torrents[infoHash])
@@ -493,7 +495,7 @@ class Server extends EventEmitter {
         socket.send(JSON.stringify({
           action: params.action === common.ACTIONS.ANNOUNCE ? 'announce' : 'scrape',
           'failure reason': err.message,
-          info_hash: common.hexToBinary(params.info_hash)
+          info_hash: hex2bin(params.info_hash)
         }), socket.onSend)
 
         this.emit('warning', err)
@@ -511,7 +513,7 @@ class Server extends EventEmitter {
           socket.infoHashes.push(params.info_hash)
         }
 
-        response.info_hash = common.hexToBinary(params.info_hash)
+        response.info_hash = hex2bin(params.info_hash)
 
         // WebSocket tracker should have a shorter interval – default: 12 seconds
         response.interval = Math.ceil(this.intervalMs / 1000 / 100)
@@ -531,8 +533,8 @@ class Server extends EventEmitter {
             action: 'announce',
             offer: params.offers[i].offer,
             offer_id: params.offers[i].offer_id,
-            peer_id: common.hexToBinary(params.peer_id),
-            info_hash: common.hexToBinary(params.info_hash)
+            peer_id: hex2bin(params.peer_id),
+            info_hash: hex2bin(params.info_hash)
           }), peer.socket.onSend)
           debug('sent offer to %s from %s', peer.peerId, params.peer_id)
         })
@@ -564,8 +566,8 @@ class Server extends EventEmitter {
             action: 'announce',
             answer: params.answer,
             offer_id: params.offer_id,
-            peer_id: common.hexToBinary(params.peer_id),
-            info_hash: common.hexToBinary(params.info_hash)
+            peer_id: hex2bin(params.peer_id),
+            info_hash: hex2bin(params.info_hash)
           }), toPeer.socket.onSend)
           debug('sent answer to %s from %s', toPeer.peerId, params.peer_id)
 
@@ -690,7 +692,7 @@ class Server extends EventEmitter {
         } else if (params.compact === 0) {
           // IPv6 peers are not separate for non-compact responses
           response.peers = response.peers.map(peer => ({
-            'peer id': common.hexToBinary(peer.peerId),
+            'peer id': hex2bin(peer.peerId),
             ip: peer.ip,
             port: peer.port
           }))
@@ -734,7 +736,7 @@ class Server extends EventEmitter {
       }
 
       results.forEach(result => {
-        response.files[common.hexToBinary(result.infoHash)] = {
+        response.files[hex2bin(result.infoHash)] = {
           complete: result.complete || 0,
           incomplete: result.incomplete || 0,
           downloaded: result.complete || 0 // TODO: this only provides a lower-bound
@@ -811,4 +813,4 @@ function toNumber (x) {
 
 function noop () {}
 
-module.exports = Server
+export default Server
