@@ -1,8 +1,8 @@
-const Client = require('../')
-const commonTest = require('./common')
-const fixtures = require('webtorrent-fixtures')
-const get = require('simple-get')
-const test = require('tape')
+import Client from '../index.js'
+import commonTest from './common.js'
+import fixtures from 'webtorrent-fixtures'
+import fetch from 'cross-fetch-ponyfill'
+import test from 'tape'
 
 const peerId = Buffer.from('-WW0091-4ea5886ce160')
 const unknownPeerId = Buffer.from('01234567890123456789')
@@ -30,89 +30,94 @@ function parseHtml (html) {
 }
 
 test('server: get empty stats', t => {
-  t.plan(11)
+  t.plan(10)
 
-  commonTest.createServer(t, 'http', (server, announceUrl) => {
+  commonTest.createServer(t, 'http', async (server, announceUrl) => {
     const url = announceUrl.replace('/announce', '/stats')
 
-    get.concat(url, (err, res, data) => {
+    let res
+    try {
+      res = await fetch(url)
+    } catch (err) {
       t.error(err)
+    }
+    const data = Buffer.from(await res.arrayBuffer())
 
-      const stats = parseHtml(data.toString())
-      t.equal(res.statusCode, 200)
-      t.equal(stats.torrents, 0)
-      t.equal(stats.activeTorrents, 0)
-      t.equal(stats.peersAll, 0)
-      t.equal(stats.peersSeederOnly, 0)
-      t.equal(stats.peersLeecherOnly, 0)
-      t.equal(stats.peersSeederAndLeecher, 0)
-      t.equal(stats.peersIPv4, 0)
-      t.equal(stats.peersIPv6, 0)
+    const stats = parseHtml(data.toString())
+    t.equal(res.status, 200)
+    t.equal(stats.torrents, 0)
+    t.equal(stats.activeTorrents, 0)
+    t.equal(stats.peersAll, 0)
+    t.equal(stats.peersSeederOnly, 0)
+    t.equal(stats.peersLeecherOnly, 0)
+    t.equal(stats.peersSeederAndLeecher, 0)
+    t.equal(stats.peersIPv4, 0)
+    t.equal(stats.peersIPv6, 0)
 
-      server.close(() => { t.pass('server closed') })
-    })
+    server.close(() => { t.pass('server closed') })
   })
 })
 
 test('server: get empty stats with json header', t => {
-  t.plan(11)
+  t.plan(10)
 
-  commonTest.createServer(t, 'http', (server, announceUrl) => {
+  commonTest.createServer(t, 'http', async (server, announceUrl) => {
     const opts = {
       url: announceUrl.replace('/announce', '/stats'),
       headers: {
         accept: 'application/json'
-      },
-      json: true
+      }
     }
-
-    get.concat(opts, (err, res, stats) => {
+    let res
+    try {
+      res = await fetch(announceUrl.replace('/announce', '/stats'), opts)
+    } catch (err) {
       t.error(err)
+    }
+    const stats = await res.json()
 
-      t.equal(res.statusCode, 200)
-      t.equal(stats.torrents, 0)
-      t.equal(stats.activeTorrents, 0)
-      t.equal(stats.peersAll, 0)
-      t.equal(stats.peersSeederOnly, 0)
-      t.equal(stats.peersLeecherOnly, 0)
-      t.equal(stats.peersSeederAndLeecher, 0)
-      t.equal(stats.peersIPv4, 0)
-      t.equal(stats.peersIPv6, 0)
+    t.equal(res.status, 200)
+    t.equal(stats.torrents, 0)
+    t.equal(stats.activeTorrents, 0)
+    t.equal(stats.peersAll, 0)
+    t.equal(stats.peersSeederOnly, 0)
+    t.equal(stats.peersLeecherOnly, 0)
+    t.equal(stats.peersSeederAndLeecher, 0)
+    t.equal(stats.peersIPv4, 0)
+    t.equal(stats.peersIPv6, 0)
 
-      server.close(() => { t.pass('server closed') })
-    })
+    server.close(() => { t.pass('server closed') })
   })
 })
 
 test('server: get empty stats on stats.json', t => {
-  t.plan(11)
+  t.plan(10)
 
-  commonTest.createServer(t, 'http', (server, announceUrl) => {
-    const opts = {
-      url: announceUrl.replace('/announce', '/stats.json'),
-      json: true
-    }
-
-    get.concat(opts, (err, res, stats) => {
+  commonTest.createServer(t, 'http', async (server, announceUrl) => {
+    let res
+    try {
+      res = await fetch(announceUrl.replace('/announce', '/stats.json'))
+    } catch (err) {
       t.error(err)
+    }
+    const stats = await res.json()
 
-      t.equal(res.statusCode, 200)
-      t.equal(stats.torrents, 0)
-      t.equal(stats.activeTorrents, 0)
-      t.equal(stats.peersAll, 0)
-      t.equal(stats.peersSeederOnly, 0)
-      t.equal(stats.peersLeecherOnly, 0)
-      t.equal(stats.peersSeederAndLeecher, 0)
-      t.equal(stats.peersIPv4, 0)
-      t.equal(stats.peersIPv6, 0)
+    t.equal(res.status, 200)
+    t.equal(stats.torrents, 0)
+    t.equal(stats.activeTorrents, 0)
+    t.equal(stats.peersAll, 0)
+    t.equal(stats.peersSeederOnly, 0)
+    t.equal(stats.peersLeecherOnly, 0)
+    t.equal(stats.peersSeederAndLeecher, 0)
+    t.equal(stats.peersIPv4, 0)
+    t.equal(stats.peersIPv6, 0)
 
-      server.close(() => { t.pass('server closed') })
-    })
+    server.close(() => { t.pass('server closed') })
   })
 })
 
 test('server: get leecher stats.json', t => {
-  t.plan(11)
+  t.plan(10)
 
   commonTest.createServer(t, 'http', (server, announceUrl) => {
     // announce a torrent to the tracker
@@ -127,33 +132,32 @@ test('server: get leecher stats.json', t => {
 
     client.start()
 
-    server.once('start', () => {
-      const opts = {
-        url: announceUrl.replace('/announce', '/stats.json'),
-        json: true
-      }
-
-      get.concat(opts, (err, res, stats) => {
+    server.once('start', async () => {
+      let res
+      try {
+        res = await fetch(announceUrl.replace('/announce', '/stats.json'))
+      } catch (err) {
         t.error(err)
+      }
+      const stats = await res.json()
 
-        t.equal(res.statusCode, 200)
-        t.equal(stats.torrents, 1)
-        t.equal(stats.activeTorrents, 1)
-        t.equal(stats.peersAll, 1)
-        t.equal(stats.peersSeederOnly, 0)
-        t.equal(stats.peersLeecherOnly, 1)
-        t.equal(stats.peersSeederAndLeecher, 0)
-        t.equal(stats.clients.WebTorrent['0.91'], 1)
+      t.equal(res.status, 200)
+      t.equal(stats.torrents, 1)
+      t.equal(stats.activeTorrents, 1)
+      t.equal(stats.peersAll, 1)
+      t.equal(stats.peersSeederOnly, 0)
+      t.equal(stats.peersLeecherOnly, 1)
+      t.equal(stats.peersSeederAndLeecher, 0)
+      t.equal(stats.clients.WebTorrent['0.91'], 1)
 
-        client.destroy(() => { t.pass('client destroyed') })
-        server.close(() => { t.pass('server closed') })
-      })
+      client.destroy(() => { t.pass('client destroyed') })
+      server.close(() => { t.pass('server closed') })
     })
   })
 })
 
 test('server: get leecher stats.json (unknown peerId)', t => {
-  t.plan(11)
+  t.plan(10)
 
   commonTest.createServer(t, 'http', (server, announceUrl) => {
     // announce a torrent to the tracker
@@ -168,27 +172,26 @@ test('server: get leecher stats.json (unknown peerId)', t => {
 
     client.start()
 
-    server.once('start', () => {
-      const opts = {
-        url: announceUrl.replace('/announce', '/stats.json'),
-        json: true
-      }
-
-      get.concat(opts, (err, res, stats) => {
+    server.once('start', async () => {
+      let res
+      try {
+        res = await fetch(announceUrl.replace('/announce', '/stats.json'))
+      } catch (err) {
         t.error(err)
+      }
+      const stats = await res.json()
 
-        t.equal(res.statusCode, 200)
-        t.equal(stats.torrents, 1)
-        t.equal(stats.activeTorrents, 1)
-        t.equal(stats.peersAll, 1)
-        t.equal(stats.peersSeederOnly, 0)
-        t.equal(stats.peersLeecherOnly, 1)
-        t.equal(stats.peersSeederAndLeecher, 0)
-        t.equal(stats.clients.unknown['01234567'], 1)
+      t.equal(res.status, 200)
+      t.equal(stats.torrents, 1)
+      t.equal(stats.activeTorrents, 1)
+      t.equal(stats.peersAll, 1)
+      t.equal(stats.peersSeederOnly, 0)
+      t.equal(stats.peersLeecherOnly, 1)
+      t.equal(stats.peersSeederAndLeecher, 0)
+      t.equal(stats.clients.unknown['01234567'], 1)
 
-        client.destroy(() => { t.pass('client destroyed') })
-        server.close(() => { t.pass('server closed') })
-      })
+      client.destroy(() => { t.pass('client destroyed') })
+      server.close(() => { t.pass('server closed') })
     })
   })
 })
